Add files via upload
Notes: 

Be sure your PDF printing program is printing at 100% scale. Many PDF programs like Acrobat or Preview default to something like 91% and that will affect the scaling. WHen printed it should be a 5mm grid. 

The graph is based on the idea of windowing the plotter work space to elimiate some fo the difficult math.  For it to match up with the bed, your program must implement the viewport and window settings indicated on the bottom of the page.